% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data Wrangling.R
\name{clean_all_column_cleaner}
\alias{clean_all_column_cleaner}
\title{Run column cleaner on all tibbles loaded in assuming from mcmicro output
See: column_cleaner for explaination}
\usage{
clean_all_column_cleaner(tibble_list)
}
\arguments{
\item{tibble_list}{list of tibble objects}
}
\value{
cleaned list of tibble objects
}
\description{
Run column cleaner on all tibbles loaded in assuming from mcmicro output
See: column_cleaner for explaination
}
\examples{
clean_all_column_cleaner(tibble_list)
}
