% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data Wrangling.R
\name{column_cleaner}
\alias{column_cleaner}
\title{Modify Column Name for CSV file assuming mcmicro HistoCat output
Takes the type of data, sample name and removes it from the column name for ease of reading and compatiability with Ajit's IMAAP
Adds column for type of data and replacing the sample id column with names
default for histocat's columns are: "Cell_[imagename]marker name"
rename columns to reflect:
ImageId = [imagename]
Column = marker name
Add column for what kind = Cell
assumption is marker ids start at location 3}
\usage{
column_cleaner(tib)
}
\arguments{
\item{data}{The tible object}
}
\value{
the modified tible object
}
\description{
Modify Column Name for CSV file assuming mcmicro HistoCat output
Takes the type of data, sample name and removes it from the column name for ease of reading and compatiability with Ajit's IMAAP
Adds column for type of data and replacing the sample id column with names
default for histocat's columns are: "Cell_[imagename]marker name"
rename columns to reflect:
ImageId = [imagename]
Column = marker name
Add column for what kind = Cell
assumption is marker ids start at location 3
}
\examples{
column_cleaner(tibbleobject)
}
