% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data Wrangling.R
\name{csv_reader}
\alias{csv_reader}
\title{Tibble CSV reader that prints what is being loaded}
\usage{
csv_reader(path)
}
\arguments{
\item{path}{is the location of the file to be loaded.}
}
\value{
tibble object
}
\description{
Tibble CSV reader that prints what is being loaded
}
\examples{
csv_reader("C:/Users/Nathan/Dropbox/@Dana Farber/CyCif/108 - ProjectID_28_21_BTIL_TNBC_RODIG/data_UNET/Parameter_Exploring/original/JENNTNBC01.csv")
}
