% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data Wrangling.R
\name{load_all_csv_files}
\alias{load_all_csv_files}
\title{Load in all CSV files assuming from mcmicro output}
\usage{
load_all_csv_files(location)
}
\arguments{
\item{location}{The path location where all .csv files are stored}
}
\value{
list of tibble objects
}
\description{
Load in all CSV files assuming from mcmicro output
}
\examples{
load_all_csv_files("C:/Users/Nathan/Dropbox/@Dana Farber/CyCif/108 - ProjectID_28_21_BTIL_TNBC_RODIG/data_UNET/Parameter_Exploring/original/")
}
